requirements.txt
fastapi
uvicorn[standard]
web3
eth_abi
eth_utils
base58
arrow
xlwt
xlrd==1.2.0
redis
psycopg2-binary